requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic-settings==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic==1.14.0
email-validator==2.2.0
